+++ protected/humhub/modules/stream/widgets/StreamViewer.php
@@ -8,8 +8,10 @@
 
 namespace humhub\modules\stream\widgets;
 
+use Codeception\Exception\ConfigurationException;
 use Yii;
 use yii\base\Exception;
+use yii\base\InvalidConfigException;
 use yii\helpers\Url;
 use humhub\modules\stream\actions\Stream;
 use humhub\modules\topic\models\Topic;
@@ -98,11 +100,12 @@ class StreamViewer extends JsWidget
 
     /**
      * @inheritdoc
+     * @throws InvalidConfigException
      */
     public function init()
     {
         if ($this->streamAction == "") {
-            throw new Exception('You need to set the streamAction attribute to use this widget!');
+            throw new InvalidConfigException('You need to set the streamAction attribute to use this widget!');
         }
 
         // Setup default messages
@@ -112,6 +115,8 @@ class StreamViewer extends JsWidget
         if ($this->messageStreamEmptyWithFilters == "") {
             $this->messageStreamEmptyWithFilters = Yii::t('ContentModule.widgets_views_stream', 'No matches with your selected filters!');
         }
+
+        parent::init();
     }
 
     public function getData()